\name{geom_efp}
\alias{geom_efp}
\title{Tissue map, line drawing with filled regions}
\usage{
  geom_efp(collection, mapping = NULL, data = NULL,
    stat = "identity", position = "identity", ...)
}
\arguments{
  \item{mapping}{ehm}

  \item{data}{the data}

  \item{stat}{a statistic, not needed?}

  \item{position}{not needed?}

  \item{...}{other params}
}
\value{
  a geom
}
\description{
  Line drawing
}
\author{
  Henning Redestig
}

