\name{theme_efp}
\alias{theme_efp}
\title{A theme for eFP-style pictograms}
\usage{
  theme_efp(base_size = 12, base_family = "")
}
\arguments{
  \item{base_size}{size for fonts}

  \item{base_family}{font}
}
\value{
  a theme
}
\description{
  Used as a hack to remove unwanted (pointless) scale
  annotations
}
\author{
  Henning Redestig
}

