% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exhibits}
\alias{exhibits}
\title{List available exhibits}
\usage{
exhibits()
}
\value{
a data frame describing the exhibits provided in this
package
}
\description{
An exhibit here is a line drawing with regions annotated as being
speciific tissues that can be used to visualize data in a
contextual map.
}
\examples{
exhibits()
}
\author{
Henning Redestig
}
